move methods for further usage across other brokerages
