+++ ToolBox/Program.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
  *
@@ -12,14 +12,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
 */
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
 using QuantConnect.Configuration;
 using QuantConnect.Logging;
 using QuantConnect.ToolBox.AlgoSeekFuturesConverter;
 using QuantConnect.ToolBox.AlgoSeekOptionsConverter;
+using QuantConnect.ToolBox.AlphaVantageDownloader;
 using QuantConnect.ToolBox.BinanceDownloader;
 using QuantConnect.ToolBox.BitfinexDownloader;
 using QuantConnect.ToolBox.CoarseUniverseGenerator;
@@ -40,9 +37,11 @@ using QuantConnect.ToolBox.QuandlBitfinexDownloader;
 using QuantConnect.ToolBox.QuantQuoteConverter;
 using QuantConnect.ToolBox.RandomDataGenerator;
 using QuantConnect.ToolBox.YahooDownloader;
-using QuantConnect.Util;
 using QuantConnect.ToolBox.ZerodhaDownloader;
-using QuantConnect.ToolBox.AlphaVantageDownloader;
+using QuantConnect.Util;
+using System;
+using System.IO;
+using static QuantConnect.Configuration.ApplicationParser;
 
 namespace QuantConnect.ToolBox
 {
@@ -80,7 +79,7 @@ namespace QuantConnect.ToolBox
                 {
                     case "zdl":
                     case "zerodhadownloader":
-                        ZerodhaDataDownloaderProgram.ZerodhaDataDownloader(tickers,market, resolution, securityType, fromDate, toDate);
+                        ZerodhaDataDownloaderProgram.ZerodhaDataDownloader(tickers, market, resolution, securityType, fromDate, toDate);
                         break;
                     case "gdaxdl":
                     case "gdaxdownloader":
@@ -137,8 +136,8 @@ namespace QuantConnect.ToolBox
                             tickers,
                             GetParameterOrExit(optionsObject, "security-type"),
                             GetParameterOrExit(optionsObject, "market"),
-                            resolution, 
-                            fromDate, 
+                            resolution,
+                            fromDate,
                             toDate);
                         break;
 
@@ -199,8 +198,8 @@ namespace QuantConnect.ToolBox
                     case "cadc":
                     case "coinapidataconverter":
                         CoinApiDataConverterProgram.CoinApiDataProgram(
-                            GetParameterOrExit(optionsObject, "date"), 
-                            GetParameterOrExit(optionsObject, "source-dir"), 
+                            GetParameterOrExit(optionsObject, "date"),
+                            GetParameterOrExit(optionsObject, "source-dir"),
                             GetParameterOrExit(optionsObject, "destination-dir"),
                             GetParameterOrDefault(optionsObject, "market", null));
                         break;
@@ -246,39 +245,5 @@ namespace QuantConnect.ToolBox
                 }
             }
         }
-
-        private static void PrintMessageAndExit(int exitCode = 0, string message = "")
-        {
-            if (!message.IsNullOrEmpty())
-            {
-                Console.WriteLine("\n" + message);
-            }
-            Console.WriteLine("\nUse the '--help' parameter for more information");
-            Console.WriteLine("Press any key to quit");
-            Console.ReadLine();
-            Environment.Exit(exitCode);
-        }
-
-        private static string GetParameterOrExit(IReadOnlyDictionary<string, object> optionsObject, string parameter)
-        {
-            if (!optionsObject.ContainsKey(parameter))
-            {
-                PrintMessageAndExit(1, "ERROR: REQUIRED parameter --" + parameter + "= is missing");
-            }
-            return optionsObject[parameter].ToString();
-        }
-
-        private static string GetParameterOrDefault(IReadOnlyDictionary<string, object> optionsObject, string parameter, string defaultValue)
-        {
-            object value;
-            if (!optionsObject.TryGetValue(parameter, out value))
-            {
-                Console.WriteLine($"'{parameter}' was not specified. Using default value: '{defaultValue}'");
-                return defaultValue;
-            }
-
-            return value.ToString();
-        }
-
     }
 }
